cicd: commit: Restrict workflow to 'push' event only
The pull request does not have a standard for its message, thus, it is
difficult to ensure some level of quality using static rules.

We've been suffering this for some time and as consequence most PR are
being marked as failing, while this was not really the case. To fix, we
restrict the workflow to 'push' event only.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>