api: firewall rules
This commits implements a flexible firewall for filtering SSH connections.
It gives a fine-grained control over which SSH connections reach the devices.

The firewall rules can deny/allow SSH connections from specific IP address
to a specific or a group of devices using a given username.

Each firewall rule has a priority number which is evaluated in ascending order
(starting with the lowest). It's makes a lot easier to manage a large number
of firewall rules.