ui: Apply eslint properties and resolve the warnings in the unit test
In the code of the unitTest is apply design pattern of the project.
Eslint has been updated and the generation warning with isVueInstance has
been deprecated. As the method is removed in the new version, to resolve the
warning, remove it. Errors for Eslint tests in DeviceIcon were fixed as
well as global require issues.