ui: Add error handling in devices and refactor the code
In the list of devices, use the computed to get devices and device
numbers and remove unused variables. The snackbar is modified to
support changes in the error variable, because the snackbar changes
the status of the variable every time it changes to true. Modify the
dashboard to suit the snackbar modifications.

Fix test to attend the modification in devices