agent: Add utmp and wtmp recording of pty sessions
This change implements recording of login activity to the utmp
and wtmp databases so that sessions are reported by standard
utilities such as "who", "last", "w" and "users".

The code uses the glibc functions pututline(3) and updwtmp(3)
to access the utmp and wtmp files but these functions are not
implemented within the musl libc included within the alpine
base used by the Docker agent build process. musl libc defines
PATH_UTMP as /dev/null/utmp and PATH_WTMP as /dev/null/wtmp and
pututline(3) and updwtmp(3) as null functions. The code runs
without error within an Alpine Docker container but has no
effect. It updates utmp and wtmp only when run natively or
inside a Docker container built with glibc support.

For this reason, this change is not suitable for merging but
may be useful as a basis for a more complete implementation,
either using https://skarnet.org/software/utmps/ as suggested
by @otavio or by porting pututline(3) and updwtmp(3) to Go and
including them within the agent code.

Comments and feedback on the code is very welcome as I'm not
a Go programmer and I'm sure the code can be improved.

Partially fixes https://github.com/shellhub-io/shellhub/issues/317