gateway: Fix /info to omit http host port from ssh host name
If a nonstandard port is used for HTTP, it is included in
ngx.var.http_host. Thus, when ":<ssh_port>" is appended, the
result is that there are two port numbers appended to the value
of the ssh endpoint in the /info api. The solution is to use
ngx.var.host rather than ngx.var.http_host for this case.