ui: Refactor the firewall rule to only make a request
When there was data in the first request, the list component was
re-rendered and made the same request.

To solve this problem, logic is added that when the first request is
made, the list component verifies that the request has already been
made and does not do it again because the data is already stored.