general: Rework installation script syntax
The installation script, using the `sh <(curl ...)` does not work on
Busybox shell. This has been detected in newer Torizon operating system.

We are now using:

  curl -sSf ... | sh

This is the same format used in `htts://rustup.rs` to install the Rust
compiler and has been in use for quite some time by them.

This partially revert the change in #638, however, the commit log and
issue didn't provide enough information to reproduce the original issue
so in case it breaks something we'll need to find a testing case for it.

Reported-by: Guilherme Fernandes <guilherme.fernandes@toradex.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>