ui: apply design pattern to tests
This commit applies the design pattern to improve test development.

Some rules that apply:

1- Declare the variable closest to the scope;
2- The order of the variable, first the ​​unit variables, then the lists;
3- Whenever you declare a variable that is a list, put a space at the beginning
and end;
4- If you have more than one situation that needs to use another wrapper, try
declaring in the test that it will be used;
5- If it is necessary to declare another wrapper, specify the name of all
wrappers;
6- Whenever possible, leave expect last in the test. If you are making changes
to the data and putting the expect in the middle of the test, it might be an
indication that you need to create another test.