ui: fix call requests on logout
First, the error occurs when it comes from the second or next namespace
and click in logout. This causes a request to occur on devices, session,
firewall rules and namespaces.

For this, some tests were added to ensure that the query is not carried
out. For this, the show variable must be false.

This closes #935.