api: change HTTP response from string to empty body
While documenting the API using OpenAPI standard, I've noticed some
details on the code, like JSON object and properties names, which are
either camel case or snake case, and how the API threat the response
in error case.

On the file that hold the code to care with namespace routing,
`nsadm.go`, has some returns what does not match with an empiric
pattern I've found in other files like `device.go` and `sshkeys.go`.

- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/sshkeys.go

Check it on `device.go`:
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go#L107
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go#L168
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go#L191
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go#L233
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/device.go#L271

Check it on `sshkeys.go`:
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/sshkeys.go#L48
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/sshkeys.go#L70
> There are also some HTTP errors response what does not match with this
pattern, but return a JSON with a body containing the error occurred.

Basing me on this empiric pattern, I've changed some HTTP errors'
responses which has returned a `string` to an empty body.

At the lines below, it is the only place, in the entire API code, what
works returning a string instead either a JSON object or an empty body,
thereat, I've guessed this could be good remove this variance.
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/nsadm.go#L113
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/nsadm.go#L144
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/nsadm.go#L169
- https://github.com/shellhub-io/shellhub/blob/2f5fb886b8b78148aecb2563a75d0c0ff321df4d/api/routes/nsadm.go#L201

The changes have made on routes below, and over the `404` and `409` HTTP
responses.
- `/namespaces/:id` on `delete` method
- `/namespaces/:id` on `put` method
- `/namespaces/:id/add` on `patch` method
- `/namespaces/:id/del` on `patch` method