api: change null JSON's responses to empty body
The `echo` returns a `200` HTTP code when the routing function return a nil,
but it is not explicit for the dev, so I have guessed, for the sake of
simplicity, change it to an empty body with the `200` status.

Another point is the responses which return a JSON, but with `null` inside the
body. I believe it could be a simply `no body` too with the original
status.