pkg: fix GeoIP country getting information
Currently, the Shellhub uses the GeoLite2-City's database from MaxMind to
collection the geolocation from device, but it also needs collect data from
the country whom connect. Unfortunately, GeoLite2-City does not contain
information about country, so download the appropriated database,
GeoLite2-Country, is needed.

This is necessary because the cloud's firewall check if a license is allowed
in that country.