pkg: clean ID when formatting user
To register a user, the client sends a user's model to register.
This user's model has a field called ID, that the API's store generate.
However, currently an API's client may send a request with an ID,
generating an unwanted behavior: API returns a 500 error, but the
user is registered. Furthermore, this inserted user's cannot be deleted
by the Shellhub's UI interface.

Cleaning the input data was the easiest way to remove this behavior.