api, pkg: remove function to get all roles and remove unused action
While the development API's token feature, I've noticed some issues
that could be problematic when a new member's role is added. When a
new role needs to be added, the developer should change many places
, what could lead to a probable error.

Changing from function to a map will decrease that possibility, no
more loop are needed and the code is simplified.

At this commit, the namespace's create action was also removed due
to uselessness. It is "useless" because no matter your role on a
namespace, the user might create a new one as itself as owner.