go: replace deprecated StandardClaims type with new RegisteredClaims
StandardClaims are a structured version of the JWT Claims Set, as
referenced at https://datatracker.ietf.org/doc/html/rfc7519#section-4. They
do not follow the specification exactly, since they were based on an
earlier draft of the specification and not updated.

The main difference is that they only support integer-based date
fields and singular audiences. This might lead to incompatibilities
with other JWT implementations.

The use of this is discouraged, instead the newer RegisteredClaims
struct should be used.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>