api: fix and improvement to API guard on namespace's route
The permission evaluation on the namespace route is different from all other
routes. In others routes, the JWT token ensure that the user is acting over
the own namespace, but on namespace's route, the namespace ID comes in the
URL, what force extra steps to validate if the user is a namespace's member
and may access the data.