pkg: add tags field to public key's model
it was agreed to add a new struct to define a filter rule for a
public key, either `hostname` of `tags`, instead only `hostname`.

when the JSON object comes with `hostname` in filter, it shouldn't
contain tags field, it isn't true, a serialization error will occur.

because of this change, the public key evaluation, a store test and
a migration have broke up, so it was necessary to fix it up.