api: block request when hostname is empty or tags is empty
As defined, when a filter hostname or tags are empty, the request
to either create or update a public key must fail, for keeping the
database cohesion.

Hostname can't be sent as an empty string because of Go's internal
questions with zero-value and the way of filter's structure is validated.
When hostname is empty, the Go's zero-value view it as a no present,
forcing, in validation, the existence of tags.