api: change handlers and routes error to package
In order to fix cyclic import in future routes tests, It was necessary
to turn routes' packages into api's packages. Furthermore, the route's
errors what were inside the routes package, it is now a package apart
called errors.
This refactor, what moves errors inside the routes package to an own
package, also must be done for services and store, but it is not within
the scope of this activity.