ssh,agent: kills executable initialized by a session
When a new connection is open through a command that doesn't return an
immediate output, and it doesn't have a Pty, the executable command wasn't
killed, running on background, even when client closes him/her connection.

To fix this, when the client doesn't send data anymore and its pipes are
closed, the executable is killed automatically, but only when the connection
doesn't have Pty.