ssh: remove custom error package
Error package was created to create custom error what contains internal
and external errors. The internal errors are those what happen on the
function and the external ones are for the end user read.

These errors, which are a composition of internal and external, are only
required when the SSH sends a response to the end user. At any other
place on code, the “normal” Go's errors fits better.

To simplify the code, during the refactoring I have created a closure
called `exit` that is responsibly to do what the custom error packaged
did, but in a simpler way: send to the end user a custom error response.

```go
exit := func(..., internal, external error) {
    // ...
}
```