api: invalidate token when user's role change
When a user login into a namespace, it got a JWT token containing
some information about it.

Unfortunately, this approach generated some problems related to
role update, when a user has its role changed by someone else.

Caching the token when a user login and cleaning this cache when
its role is updated, it helps to invalid a token after a role
update.