agent: fix forbidden output produced by SFTP subsystem initialization
At this commit I fixed a bug related to SFTP initialization. When the init
function inform what loglevel is set, it violates the rule what said that
the service initialized by SFTP subsystem on server cannot print on default
output.

I solved this calling the function to set the loglevel only when `agent` and
`info` subcommands are invoked by the agent's binary. `agent` subcommand is
the default one called by this binary.