cli, bin: refactor CLI commands basing it on resources
Refactoring CLI commands to work as a "resource command parameter"
pattern. All commands are, now, based on the resource it works on.

Each resource has commands related to it. A user can be created,
has its password edited and remove; a namespace can be created and
deleted; a member can be added or removed from a namespace.

To interact with the new commands on the CLI service, a new script
called "cli" on bin/ folder was created. It just runs a
`docker-compose exec` passing the parameters to CLI service and
return its output to who executed the script.

The old CLI commands and scripts to interact to CLI service still
are on bin/ folder, but was tagged as deprecated, what makes it
does not appear on helper or usage command.