cli: call method to delete namespace and add tenant to usage
At the deprecated command `add-namespace` the user could pass the
`TENNANT_id` as a shell's variable to set the namespace's tenant
instead of generate a new one. Unfortunately, the new `cli` script
no more pass this variable to the cli's container, requiring the
tenant to be  defined on the own create namespace's command.

```sh
./bin/cli namespace create shellhub shellhubspace
00000000-0000-4000-0000-000000000000
```