env: remove empty quotes to represent empty values
The issue

"I have successfully followed the setup instructions for deploying the self
hosted shellhub service (v0.10.9) locally on my Ubuntu desktop. I have created
my first user and namespace, but every time I try and login to the service
(with my credentials) I get a small dialog popping up saying "The request has
failed. Please try again". It does this even if I type in rubbish into the
username and password."

The response

"@drumwhizz the problem seems to be with the *docker-compose* version what you
are using.

I'm assuming you are using the Ubuntu 20.04, sorry.

Looking for the *docker-compose* package on your distro, I've checked that
this version, 1.25.0, doesn't support empty "" string as null on .env.

To work around that issue, you can run that script. It'll remove all `""`
from the `.env`, allowing the API server to run normally.

sed -i 's,"",,g' .env

The explanation

When using docker-compose in version > 1.25.0, Sentry's environment variable is
not empty when its value on the .env is SHELLHUB_SENTRY_DSN="", but
contains the double quotes as a string itself.

https://i.imgur.com/5RQfADu.png

On the < 1.26.0, the docker-compose has delegated to `python-dotenv` the
processing of `.env` file, what becoming the `SHELLHUB_SENTRY_DSN=""` valid as
empty environmental variable on the container.

https://i.imgur.com/ZjLRl97.png

The PR

To avoid this problem again, it was requested to remove the attribution of
empty string as VALUE="" on .env file.