api,pkg: add event enqueue and handlers for some session events
Here, I have added a way to detect events on the API.

When the service response with a successful response to the route who call it,
it will push an event to the event's queue to be managed by the worker.

Currently, the worker gets all events, what are pushed when session is created,
finished and authenticated successfully, and save all of them into a collection
in MongoDB database called "events".

The basic structure of this collection are:
- A field called type, what is the type of the event. Currently, the type
begins with `event:` followed by the resource and the action, separated by `:`.
For instance: `event:session:created`.
- A field called payload with the request data received, the response data
returned and the date of when the event was pushed.
- A field called date, what represent the date of when the worker processed the
event.