agent: fix issue with results not always reaching SSH client
Previously, when running a command on an agent through the server,
the result of the command did not always reach the SSH client.
After investigation, it was suspected that a race condition
due to intermittent results. But we are fixed all the race
conditions found and the issue persisted.

During the debugging sessions, we noticed that some bytes of the
connection between the server and the agent were not arriving
as they should. This led us to believe that someone might be
interfering with the communication by reading the bytes before
us and prematurely closing the socket, as if something
was "hijacking" the socket.

To address this, a hijack was added to the HTTP connection in
reverse listener handler. With this addition, the agent's now
properly handles the connection and prevents any output loss.

Closes: #2509