gateway: update proxy_pass config to use var for upstream server
This commit updates the `proxy_pass` configuration to use a
variable for the upstram server instead of a hardcoded value.

By using a variable in the `proxy_pass` configuration, the DNS
resolution occurs dynamically rather than when nginx starts up.
This will prevent nginx from exiting with a
"host not found in upstream" error if the host is not yet
available on the network.