env: Add SHELLHUB_PUBLIC_URL_DOMAIN env
This commit adds the `SHELLHUB_PUBLIC_URL_DOMAIN` environment variable
to the `.env` file, which is used to define the public URL domain
for accessing devices via HTTP in a ShellHub instance.
The value of this variable is used in the `gateway/shellhub.conf`
file to generate the public URL for HTTP access.

The `docker-compose.yml` file is also updated to pass this variable
to the gateway container.

NOTE: If `SHELLHUB_PUBLIC_URL_DOMAIN` is not defined,
the value of `SHELLHUB_DOMAIN` is used as the default value.