api,pkg: add slot strategy to block device switching
Here I introduce the concept of "slot", which is a way to control the number of
devices that namespace can have in a given time and when there is a device
limitation. The "slot" is a document that is created when a device is
"accepted" and its status change to "removed" when the device is deleted.
A "slot" with the status "removed" expires after a period of time, allowing a
new device to be accepted at its place.