api,pkg: add system to avoid device switching
Here I introduce the concept of "removed_device", which is a way to control the
number of devices that a namespace can have in a given time and when there is a
device limitation. The "removed_device" collection contains documents that are
created when a device is deleted.

When the user tries to accept a device, we check if there is a removed device
with the same UID, if there is, we block the device from being accepted. If
there is no removed device, we check if the namespace has reached the device
limit and if it has, we check if the number of devices in the namespace is
greater than the limit. If it is, we block the device from being accepted too.

When the namespace does not have a device limit, we skip this validation.