api: change device list behavior when listing pending
We changed the behavior of device list when status is pending to list only
acceptable devices. A device can be accepted if the number of devices in a
namespace plus the number of those were removed are less than max number of
devices a namespace can have.

```
number of devices + removed devices >= max number of them onto namespace
```