api: add acceptable attribute to device and when listing it
When listing devices, we needed to identify whether they can be acceptable or
not. All devices are acceptable unless the namespace's limit is reached. Its
limit is based on how many devices the namespace has, plus the number of devices
removed between 720 hours or 30 days, being less or equal than the namespace's
max devices number.

Once this limit is reached and the device listing request is for those devices
with status "pending" or "rejected", we set "acceptable" only for devices that
were removed.