refactor(ui): remove unnecessary store and hooks related to mobile detection
* Removed a complex store that was used to determine if the browser
is mobile or not, which involved a hardcoded window width check.
Instead, I integrated the `useDisplay` composable from Vue to accurately
determine if the screen size is "lgAndUp" to display the sidebar accordingly.
This approach is more robust and reliable compared to relying on a hardcoded
value.

* Removed a bunch of hooks that were triggered by the browser's
onResize event to track window resizing. I also eliminated store-related code
related to the sidebar's state, which was primarily used for toggling the
sidebar when in mobile view.

Additionally, I made the `<AppLayout>`, pass the sidebar's visibility model
to the `<AppBar>` component using the new `defineModel` feature of Vue 3.3.

As part of the refactoring, I also converted several components to use
`<script setup>`, which offers a much cleaner code structure compared
to defineComponent usage style.