fix(gateway): pass request headers to install script route
Due to the change in install script route, what was contained inside
the nginx script, to an Echo route, the headers were removed from
NGINX configuration file, but it were not sent to the route itself.

To fix that, I had to pass these values through the NGINX file to
the API server, what was not made.

Refs: #2858