feat(agent,pkg): split "agent" binary into a package
we had a binary that is generated from the source code located in the
"agent" directory. However, we would like to convert this agent into a
package so that it can be used outside the binary as well. This was made
to enable us to create an instance of the agent without the need to
execute a binary.

REFS: #2930