feat(api): This commit is to implement 2fat validation,
we made changes to following system functions:
AuthUser/AuthUserInfo: We implemented the return of MFA status.
AuthGetToken: It adds MFA validation data to the jwt.Token.
AuthRequest: Using the data received from the token,
it adds MFA validation data to the req.Header.
AuthMFA: This is or middleware for MFA based on
the MFA status validation.
And implement five new store func for MFA.