fix(agent,revdial): allow tunnel close when IO error occurs
When the agent loses its connection with the ShellHub SSH, the goroutine
responsible for reading the "keep-alive" message returns, letting the father
function still running. This block makes the tunnel take a long time until it
stops listening due to some error, and delays the agent's reconnection to the
server.

To solve it, we send a message indicating that the reading or writing returning,
returning the father function and allowing the tunnel to restart and
reconnecting back to the SSH server.