fix(api): Ensure token is cached when getting it from service
When trying to log in through the cloud using 'user actions', the
token generated for login is invalid because it is not saved in the
cache. To resolve this issue, we need to save the token in the
`AuthGetToken` function, so that a valid token is provided on the
user actions screen.