chore(envs): change envs parser backend due unmaintained project
Currently, we are using the `kelseyhightower/envconfig` package, what it
seems to be no longer actively maintained, so we changed to
`sethvargo/go-envconfig`, what contains the same behavior as the before
package.

BREAKING CHANGE: `sethvargo/go-envconfig` is case-sensitive for the
variable names, the prefix option and required tag work differently, and
the default value is inside the env tag.

Closes #3204