fix(ssh): local port forwarding with no request type
When a client connects to the SSH server using an option that sets the
request type to `none`, the Local Port Forward doesn't know which device
to connect because the SSH Handler wasn't called.

To allow it, we have dialed to the device through the tunnel and
authenticated it. In order to avoid a lot of duplication, we also have
refactored how the SSH server creates session, and added a new kind of
SSH session what doesn't depend on SSH request type.

Resolves #3233