refactor(agent): add modes to agent
The Agent can be executed in two different modes: `Host` and `Connector`. The
`Host` mode is the default one, where the agent will listen for incoming
connections and use the host device as source of any information needed to start
itself. When running in `Connector` mode, it uses the Docker engine as this
source.