refactor(connman): restructure ConnectionManager and introduce SyncSliceMap
This commit refactors the `ConnectionManager` by replacing the `sync.Map`
for dialers with a `SyncSliceMap`.

The `SyncSliceMap` struct provides methods for loading, storing, and
deleting values in a synchronized map with slices.

The use of `SyncSliceMap` improves the management of reverse dialers and
handles multiple connections for the same identifier.

This change potentially addresses the SSH connection issue where multiple
agents using the same identifier prevented SSH connection, even when the
web interface showed them as online.