fix(ssh): fix session authentication issues in older agent versions
Versions earlier than 0.6.0 do not validate the user when receiving a public key
authentication request. This implies that requests with invalid users are
treated as "authenticated" because the connection does not raise any error.
Moreover, the agent panics after the connection ends. To avoid this, connections
with public key are not permitted when agent version is 0.5.x or earlier