chore(api,cli,pkg): remove deprecated function on validator package
As password hashing was too tight to deprecated functions, It was
necessary to refactor it in the same commit to maintain everything
working as expected. For that, I have added one new field on
`models.UserPassword`: PlainPassword.

The `PlainPassword` field contains the password as plain text, what
should only be true in creation or update services. In all other cases,
only the `HashedPassword` field should contain data. It does not affect
how the data is stored on retrieved from the database.