refactor(api,pkg): refactoring EditNamespace
Introducing a new method, `NamespaceEdit`, within the store to
facilitate the updating of a namespace's attributes. This new method
replaces the deprecated `NamespaceRename` and `NamespaceSetSessionRecord`.

Accompanying this change is the introduction of a new struct,
`NamespaceChanges`, designed to be a parameter for the `NamespaceEdit`
method. This struct consolidates all possible attributes that users can
modify. The use of the tag `bson:"omitempty"` helps avoid unnecessary
and undesirable `if` statements by handling empty values appropriately.

In the service layer, the method `EditNamespace` now calls the store
method with the specified changes. This method now takes the entire
request and determines its course of action (in this case, creating the
`changes` struct). The validation process within the service is removed
to centralize it in the router, streamlining code complexity.

As part of deprecating `NamespaceRename`, we're also renaming the
permission to `Namespace.Update` and `NamespaceUpdate`.