feat(api,ssh,pkg): adding namespace's connection announcement
The connection announcement of a namespace is a custom UTF-8 valid
string provided by the user, with a maximum length of 127 characters,
which which can be written in a SSH session.

The SSH server must decide when to write this message in an established
connection and must skip this stage if the connection announcement is an
empty string.

A user can set the connection announcement by making a PUT request to
the existing `/api/namespaces/:tenant` route. The announcement must
leave under's `settings.connection_announcement` body:

A user can set the connection announcement by making a PUT request to
the existing `/api/namespaces/:tenant` route. The announcement must be
placed under the `settings.connection_announcement` field:

```json
{
    "settings": {
        "connection_announcement": "My awesome connection announcement"
    }
}
```

> NOTE: An empty string can be sent in this request to disable this
feature.

To retrieve the announcement, users can make a GET request to
`api/namespaces/:tenant`, where the announcement will be available in
`settings.connection_announcement`.

To retrieve the announcement, users can make a GET request to
`api/namespaces/:tenant`, where the announcement will be available in
`settings.connection_announcement`.

There are also two important things to notice:

1. A new method `NamespaceLookup` is introduced in the internal client,
   which can help the SSH session retrieve the connection announcement.
   The server must decide when to write this message in the session.

2. A new migration with ID 64 is added to automatically set the connection
   announcement to an empty string for already existing namespaces.