fix(api): set right acceptable value when listing devices without status
When listing devices with a non empty status, i.g `accepted`, `pending`,
the `acceptable` field was right defined. However, if it is set to
empty, no logic was applied, setting the field to `false`, the default
value in Go.

To solve the issue, we evaluate the `acceptable` field based on devices
in `removed_devices`, for cloud instances, its `accepted` status when it
is already `accepted` and when the namespace has a limit defined.

Fixes: #3527