chore(api,gateway): bind a random port on api startup
Introduces a new `SHELLHUB_API_PORT` variable, which allows the API to
randomly select a port when it starts up.

While binding the API to port 8080 might seem convenient, it can lead to
collision errors when running multiple Docker Compose instances (e.g.,
testcontainers) because each instance shares the same network.

To illustrate, consider two separate Docker Compose instances labeled
"A" and "B". In "A", a gateway is running on port 11111, and in "B", a
gateway is running on port 22222. Since both APIs are running on port
8080 and both instances share the same network, gateway "A" might
mistakenly request data from "B", and vice versa.

> !Note: Due to the use of internalclient, this environment should be
  sent to other services such as "ssh".