refactor: move dev agent service to it own compose file
We moved the Agent's compose service to a dedicate compose file to
allow more flexibility when running the dev environment, especially for
the integration tests.

The current behavior hasn't changed; when the `SHELLHUB_ENV` is set to
development, the Agent still starts together with the services, but now
we could separate this process using a env variable more easily in the
future.