refactor(ssh): use auth handlers to start SSH connection
The SSH's ShellHub server is the gateway between the user and its
device. Each device added to ShellHub, connects to the gateway, allowing
ShellHub's SSH server to be it manager. When a client wants to connect
to them device, this request comes first into our SSH server and, after
ward, to the device itself.

As our server is also an SSH server, what makes the SSH client
understands that the user first connects to a server and is redirected
to another one, causing problems with some SSH clients' authentication
detection.

In order to solve it, we, now, connect to the device before starting the
session itself, what makes this redirect seamless to the SSH clients.

Refs to #3240