refactor(ssh): remove unnecessary handling for session channel
In our SSH server, we have had to replicate the request made to it to
the agent, to save the session information in the API or record the
session when it is enabled. However, we have noticed that it could be
simplified using the SSH's channels, redirecting data from the channel
created by the client to the agent directly.